+++ projects/js-packages/publicize-components/src/components/connection-management/connection-status.tsx
@@ -21,19 +21,15 @@ export function ConnectionStatus( {
 	service,
 	onConfirmReconnect,
 }: ConnectionStatusProps ) {
-	if ( connection.status === undefined || connection.status === 'ok' ) {
+	if ( connection.status !== 'broken' ) {
 		return null;
-	}
-
-	let notice = __( 'There is an issue with this connection.', 'jetpack' );
-
-	if ( connection.status === 'refresh-failed' ) {
-		notice = __( 'The connection seems to have expired.', 'jetpack' );
 	}
 
 	return (
 		<div>
-			<span className="description">{ notice }</span>
+			<span className="description">
+				{ __( 'There is an issue with this connection.', 'jetpack' ) }
+			</span>
 			&nbsp;
 			<Reconnect
 				connection={ connection }