+++ projects/js-packages/publicize-components/src/components/services/service-connection-info.tsx
@@ -1,4 +1,4 @@
-import { IconTooltip } from '@automattic/jetpack-components';
+import { IconTooltip, Text } from '@automattic/jetpack-components';
 import { __ } from '@wordpress/i18n';
 import { Connection } from '../../social-store/types';
 import { ConnectionName } from '../connection-management/connection-name';
@@ -11,9 +11,14 @@ import { SupportedService } from './use-supported-services';
 export type ServiceConnectionInfoProps = {
 	connection: Connection;
 	service: SupportedService;
+	isAdmin?: boolean;
 };
 
-export const ServiceConnectionInfo = ( { connection, service }: ServiceConnectionInfoProps ) => {
+export const ServiceConnectionInfo = ( {
+	connection,
+	service,
+	isAdmin,
+}: ServiceConnectionInfoProps ) => {
 	return (
 		<div className={ styles[ 'service-connection' ] }>
 			<div>
@@ -29,19 +34,31 @@ export const ServiceConnectionInfo = ( { connection, service }: ServiceConnectio
 			</div>
 			<div className={ styles[ 'connection-details' ] }>
 				<ConnectionName connection={ connection } />
-				{ connection.status === 'broken' ? (
-					<ConnectionStatus connection={ connection } service={ service } />
-				) : (
-					<div className={ styles[ 'mark-shared-wrap' ] }>
-						<MarkAsShared connection={ connection } />
-						<IconTooltip placement="top" inline={ false } shift>
-							{ __(
-								'If enabled, the connection will be available to all administrators, editors, and authors.',
-								'jetpack'
-							) }
-						</IconTooltip>
-					</div>
-				) }
+				{ ( conn => {
+					if ( conn.status === 'broken' ) {
+						return <ConnectionStatus connection={ conn } service={ service } />;
+					}
+
+					if ( isAdmin ) {
+						return (
+							<div className={ styles[ 'mark-shared-wrap' ] }>
+								<MarkAsShared connection={ conn } />
+								<IconTooltip placement="top" inline={ false } shift>
+									{ __(
+										'If enabled, the connection will be available to all administrators, editors, and authors.',
+										'jetpack'
+									) }
+								</IconTooltip>
+							</div>
+						);
+					}
+
+					return ! conn.can_disconnect ? (
+						<Text className={ styles.description }>
+							{ __( 'This connection is added by a site administrator.', 'jetpack' ) }
+						</Text>
+					) : null;
+				} )( connection ) }
 			</div>
 			<div className={ styles[ 'connection-actions' ] }>
 				<Disconnect