+++ projects/packages/identity-crisis/.phan/baseline.php
@@ -8,25 +8,9 @@
  * (can be combined with --load-baseline)
  */
 return [
-    // # Issue statistics:
-    // PhanTypeMismatchArgument : 5 occurrences
-    // PhanNoopNew : 1 occurrence
-    // PhanParamTooMany : 1 occurrence
-    // PhanPluginRedundantAssignment : 1 occurrence
-    // PhanRedundantCondition : 1 occurrence
-    // PhanTypeMismatchArgumentInternal : 1 occurrence
-    // PhanTypeMismatchArgumentProbablyReal : 1 occurrence
-    // PhanTypeMismatchProperty : 1 occurrence
-    // PhanTypeMismatchReturnNullable : 1 occurrence
-
+    // This baseline has no suppressions
     // Currently, file_suppressions and directory_suppressions are the only supported suppressions
     'file_suppressions' => [
-        'src/class-identity-crisis.php' => ['PhanPluginRedundantAssignment', 'PhanRedundantCondition'],
-        'src/class-rest-endpoints.php' => ['PhanTypeMismatchArgumentProbablyReal'],
-        'src/class-url-secret.php' => ['PhanTypeMismatchProperty', 'PhanTypeMismatchReturnNullable'],
-        'tests/php/test-identity-crisis.php' => ['PhanParamTooMany', 'PhanTypeMismatchArgument'],
-        'tests/php/test-rest-endpoints.php' => ['PhanNoopNew'],
-        'tests/php/test-url-secret.php' => ['PhanTypeMismatchArgumentInternal'],
     ],
     // 'directory_suppressions' => ['src/directory_name' => ['PhanIssueName1', 'PhanIssueName2']] can be manually added if needed.
     // (directory_suppressions will currently be ignored by subsequent calls to --save-baseline, but may be preserved in future Phan releases)