+++ projects/packages/identity-crisis/src/class-exception.php
@@ -1,13 +1,10 @@
-<?php
+<?php // phpcs:ignore
 /**
  * Exception class for the Identity Crisis package.
  *
  * @package  automattic/jetpack-identity-crisis
+ * @deprecated 0.21.0
+ * @see \Automattic\Jetpack\IdentityCrisis\Exception
  */
 
-namespace Automattic\Jetpack\IdentityCrisis;
-
-/**
- * Exception class for the Identity Crisis package.
- */
-class Exception extends \Exception {}
+require_once __DIR__ . '/../../jetpack-connection/src/identity-crisis/class-exception.php';