+++ projects/plugins/automattic-for-agencies-client/src/js/components/connection-card/index.jsx
@@ -60,7 +60,7 @@ function ConnectionContent( { onShowSharingDetailsClick } ) {
 				<div className={ styles[ 'terms-of-service' ] }>
 					{ createInterpolateElement(
 						__(
-							'By clicking the <strong>connect this site</strong> button, you agree to our <tosLink>Terms of Service</tosLink> and to <shareDetailsLink>share details</shareDetailsLink> with WordPress.com.',
+							'By clicking <strong>connect this site</strong>, you agree to our <tosLink>Terms of Service</tosLink> and to <shareDetailsLink>sync your site‘s data</shareDetailsLink> with us.',
 							'automattic-for-agencies-client'
 						),
 						{
@@ -99,7 +99,7 @@ function ConnectionContent( { onShowSharingDetailsClick } ) {
 
 /**
  * Detail Sharing Content
- * The copy for the "share details" information linked from the TOS.
+ * The copy for the "sync your site‘s data" information linked from the TOS.
  *
  * @param {object}   props                             - Component props
  * @param {Function} props.onCloseSharingDetailsClick  - Callback to close the sharing details.
@@ -121,7 +121,7 @@ function DetailSharingContent( { onCloseSharingDetailsClick } ) {
 			<div className={ styles.card__prose }>
 				<h2>
 					{ __(
-						'What data is shared from your site with WordPress.com',
+						'What data is synced between your site and WordPress.com',
 						'automattic-for-agencies-client'
 					) }
 				</h2>