+++ projects/plugins/boost/app/lib/critical-css/source-providers/providers/Archive_Provider.php
@@ -105,6 +105,13 @@ class Archive_Provider extends Provider {
 			$provider_post_types[ $post_type->name ] = $post_type->name;
 		}
 
+		/**
+		 * Filters the post types used for Critical CSS
+		 *
+		 * @param array $post_types The array of post types to be used
+		 *
+		 * @since   1.0.0
+		 */
 		return apply_filters(
 			'jetpack_boost_critical_css_post_types_archives',
 			apply_filters_deprecated(