+++ projects/plugins/jetpack/extensions/blocks/ai-assistant/inline-extensions/components/ai-assistant-input/index.tsx
@@ -18,9 +18,10 @@ import './style.scss';
  */
 import type { ExtendedInlineBlockProp } from '../../../extensions/ai-assistant';
 import type { RequestingErrorProps, RequestingStateProp } from '@automattic/jetpack-ai-client';
-import type { ReactElement, MouseEvent } from 'react';
+import type { ReactElement } from 'react';
 
 export type AiAssistantInputProps = {
+	className?: string;
 	requestingState: RequestingStateProp;
 	requestingError?: RequestingErrorProps;
 	inputRef?: React.MutableRefObject< HTMLInputElement | null >;
@@ -34,12 +35,13 @@ export type AiAssistantInputProps = {
 	tryAgain?: () => void;
 };
 
-const className = classNames(
+const defaultClassNames = classNames(
 	'jetpack-ai-assistant-extension-ai-input',
 	'wp-block' // Some themes, like Twenty Twenty, use this class to set the element's side margins.
 );
 
 export default function AiAssistantInput( {
+	className,
 	requestingState,
 	requestingError,
 	inputRef,
@@ -54,7 +56,7 @@ export default function AiAssistantInput( {
 }: AiAssistantInputProps ): ReactElement {
 	const [ value, setValue ] = useState( '' );
 	const [ placeholder, setPlaceholder ] = useState( __( 'Ask Jetpack AI to edit…', 'jetpack' ) );
-	const { autosaveAndRedirect } = useAICheckout();
+	const { checkoutUrl } = useAICheckout();
 	const { tracks } = useAnalytics();
 	const [ requestsRemaining, setRequestsRemaining ] = useState( 0 );
 	const [ showUpgradeMessage, setShowUpgradeMessage ] = useState( false );
@@ -100,18 +102,13 @@ export default function AiAssistantInput( {
 		undo?.();
 	}, [ blockType, tracks, undo ] );
 
-	const handleUpgrade = useCallback(
-		( event: MouseEvent< HTMLButtonElement > ) => {
-			tracks.recordEvent( 'jetpack_ai_upgrade_button', {
-				current_tier_slug: currentTier?.slug,
-				requests_count: requestsCount,
-				placement: 'jetpack_ai_assistant_extension',
-			} );
-
-			autosaveAndRedirect( event );
-		},
-		[ autosaveAndRedirect, currentTier?.slug, requestsCount, tracks ]
-	);
+	const handleUpgrade = useCallback( () => {
+		tracks.recordEvent( 'jetpack_ai_upgrade_button', {
+			current_tier_slug: currentTier?.slug,
+			requests_count: requestsCount,
+			placement: 'jetpack_ai_assistant_extension',
+		} );
+	}, [ currentTier?.slug, requestsCount, tracks ] );
 
 	const handleTryAgain = useCallback( () => {
 		tracks.recordEvent( 'jetpack_ai_assistant_try_again', {
@@ -156,7 +153,7 @@ export default function AiAssistantInput( {
 
 	return (
 		<ExtensionAIControl
-			className={ className }
+			className={ classNames( defaultClassNames, className ) }
 			placeholder={ placeholder }
 			disabled={ disabled }
 			value={ value }
@@ -165,6 +162,7 @@ export default function AiAssistantInput( {
 			error={ requestingError }
 			requestsRemaining={ requestsRemaining }
 			showUpgradeMessage={ showUpgradeMessage }
+			upgradeUrl={ checkoutUrl }
 			onChange={ setValue }
 			onSend={ handleSend }
 			onStop={ handleStopSuggestion }