+++ projects/plugins/jetpack/extensions/blocks/ai-assistant/inline-extensions/components/ai-assistant-toolbar-dropdown/index.tsx
@@ -10,11 +10,17 @@ import React, { useCallback } from 'react';
  * Internal dependencies
  */
 import AiAssistantToolbarDropdownContent from '../../../components/ai-assistant-toolbar-dropdown/dropdown-content';
+import useTransformToAssistant from '../../../hooks/use-transform-to-assistant';
 /*
  * Types
  */
-import type { OnRequestSuggestion } from '../../../components/ai-assistant-toolbar-dropdown/dropdown-content';
+import type {
+	AiAssistantDropdownOnChangeOptionsArgProps,
+	OnRequestSuggestion,
+} from '../../../components/ai-assistant-toolbar-dropdown/dropdown-content';
 import type { ExtendedInlineBlockProp } from '../../../extensions/ai-assistant';
+import type { PromptTypeProp } from '../../../lib/prompt';
+import type { BlockBehavior } from '../../types';
 import type { ReactElement } from 'react';
 
 type AiAssistantExtensionToolbarDropdownContentProps = {
@@ -35,14 +41,51 @@ function AiAssistantExtensionToolbarDropdownContent( {
 	onAskAiAssistant,
 	onRequestSuggestion,
 }: AiAssistantExtensionToolbarDropdownContentProps ) {
-	const handleRequestSuggestion: OnRequestSuggestion = ( ...args ) => {
-		onRequestSuggestion?.( ...args );
-		onClose?.();
+	const { canTransformToAIAssistant, transformToAIAssistant, getSelectedBlockClientIds } =
+		useTransformToAssistant();
+
+	const handleToolbarButtonClick = useCallback(
+		( request?: {
+			promptType: PromptTypeProp;
+			options?: AiAssistantDropdownOnChangeOptionsArgProps;
+			humanText?: string;
+		} ) => {
+			const selectedBlockIds = getSelectedBlockClientIds();
+			const [ clientId ] = selectedBlockIds;
+
+			if (
+				selectedBlockIds.length < 2 ||
+				! canTransformToAIAssistant( { clientId, blockName: blockType } )
+			) {
+				// If there is only one selected block or the block cannot be transformed, proceed to open the extension input.
+				if ( request ) {
+					onRequestSuggestion?.( request.promptType, request.options, request.humanText );
+				} else {
+					onAskAiAssistant?.();
+				}
+				onClose?.();
+			} else {
+				// If there are multiple blocks selected, replace them with a single AI Assistant block.
+				transformToAIAssistant( { request } );
+			}
+		},
+		[
+			blockType,
+			canTransformToAIAssistant,
+			getSelectedBlockClientIds,
+			onAskAiAssistant,
+			onClose,
+			onRequestSuggestion,
+			transformToAIAssistant,
+		]
+	);
+
+	const handleRequestSuggestion: OnRequestSuggestion = ( promptType, options, humanText ) => {
+		handleToolbarButtonClick( { promptType, options, humanText } );
 	};
 
-	const handleAskAiAssistant = () => {
-		onAskAiAssistant?.();
-		onClose?.();
+	const handleAskAiAssistant = async () => {
+		handleToolbarButtonClick();
 	};
 
 	return (
@@ -56,6 +99,7 @@ function AiAssistantExtensionToolbarDropdownContent( {
 }
 
 type AiAssistantExtensionToolbarDropdownProps = {
+	behavior: BlockBehavior;
 	blockType: ExtendedInlineBlockProp;
 	label?: string;
 	onAskAiAssistant: () => void;
@@ -63,6 +107,7 @@ type AiAssistantExtensionToolbarDropdownProps = {
 };
 
 export default function AiAssistantExtensionToolbarDropdown( {
+	behavior,
 	blockType,
 	label = __( 'AI Assistant', 'jetpack' ),
 	onAskAiAssistant,
@@ -107,11 +152,22 @@ export default function AiAssistantExtensionToolbarDropdown( {
 				variant: 'toolbar',
 			} }
 			renderToggle={ ( { isOpen, onToggle } ) => {
+				const handleClick = () => {
+					switch ( behavior ) {
+						case 'action':
+							handleAskAiAssistant();
+							break;
+						case 'dropdown':
+							onToggle();
+							break;
+					}
+				};
+
 				return (
 					<ToolbarButton
 						className="jetpack-ai-assistant__button"
 						showTooltip
-						onClick={ onToggle }
+						onClick={ handleClick }
 						aria-haspopup="true"
 						aria-expanded={ isOpen }
 						label={ label }