+++ projects/js-packages/publicize-components/CHANGELOG.md
@@ -5,6 +5,34 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.53.0] - 2024-05-29
+### Added
+- Added tests for manage connections modal and services [#37582]
+
+### Removed
+- Social | Removed sharing buttons info from connections modal [#37593]
+
+### Fixed
+- Social | Hide "Mark as shared" for non-admin authors [#37595]
+
+## [0.52.0] - 2024-05-27
+### Added
+- Added tests. [#37516]
+- Added the connection modal to the editor. [#37405]
+
+### Changed
+- Refactored the connection list UI. [#37437]
+- Social | Cleaned up connections management list. [#37570]
+- Social | Fixed no connections UI for editor. [#37571]
+- Social | Improved reconnect flow to automatically open the connection window. [#37467]
+- Social | Made the editor connections modal behavior similar to the modal on the social admin page. [#37501]
+- Social | Updated connection fields to match the API schema. [#37529]
+- Social | Updated the connection test results endpoint for front-end. [#37531]
+
+### Fixed
+- Fixed an issue where non-admins saw action on connection management. [#37507]
+- Fixed UI bug with border. [#37490]
+
 ## [0.51.1] - 2024-05-20
 ### Added
 - Added globalize/unglobalize connection UI. [#37377]
@@ -684,6 +712,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Updated package dependencies. [#24470]
 
+[0.53.0]: https://github.com/Automattic/jetpack-publicize-components/compare/v0.52.0...v0.53.0
+[0.52.0]: https://github.com/Automattic/jetpack-publicize-components/compare/v0.51.1...v0.52.0
 [0.51.1]: https://github.com/Automattic/jetpack-publicize-components/compare/v0.51.0...v0.51.1
 [0.51.0]: https://github.com/Automattic/jetpack-publicize-components/compare/v0.50.0...v0.51.0
 [0.50.0]: https://github.com/Automattic/jetpack-publicize-components/compare/v0.49.3...v0.50.0