+++ projects/packages/options/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.1] - 2024-05-30
+### Changed
+- Internal updates.
+
 ## [2.0.0] - 2024-02-07
 ### Changed
 - The package now requires PHP >= 7.0. [#34192]
@@ -195,6 +199,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
+[2.0.1]: https://github.com/Automattic/jetpack-options/compare/v2.0.0...v2.0.1
 [2.0.0]: https://github.com/Automattic/jetpack-options/compare/v1.16.6...v2.0.0
 [1.16.6]: https://github.com/Automattic/jetpack-options/compare/v1.16.5...v1.16.6
 [1.16.5]: https://github.com/Automattic/jetpack-options/compare/v1.16.4...v1.16.5